changed amount and id data type

+++ domain/entity/balance.go
@@ -1,9 +1,9 @@
 package entity
 
 type Balance struct {
-	UserID int     `json:"user_id"`
-	Amount float64 `json:"amount"`
-	TypeOp int     `json:"type_op,omitempty"`
+	UserID int64 `json:"user_id"`
+	Amount int64 `json:"amount"`
+	TypeOp int   `json:"type_op,omitempty"`
 }
 
 const (
